packages/embers/pyproject.toml
[project]
dependencies = [
    "base58 (>=2.1,<3.0.0)",
    "betterproto2 (>=0.9,<0.10.0)",
    "crc (>=7.1,<8.0.0)",
    "ecdsa (>=0.19,<0.20.0)",
    "pytest (>=9.0,<10.0.0)",
    "python-zbase32 (>=0.1,<0.2.0)",
    "requests (>=2.32,<3.0.0)",
    "websocket-client (>=1.9,<2.0.0)",
]
name = "embers-tests"
requires-python = ">=3.13,<4"
version = "0.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires      = ["poetry-core>=2.0.0,<3.0.0"]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.76"
ruff         = "^0.14"

[tool.ruff]
exclude        = ["tests/protobuf"]
include        = ["tests/*.py"]
line-length    = 120
target-version = "py313"

[tool.ruff.lint]
ignore = [
    "ANN401",  # typing.Any
    "D",       # docstrings
    "PLR0913", # Too many arguments
    "PLR2004", # magic value
    "S101",    # assert
    "S311",    # suspicious-non-cryptographic-random-usage
    "TC001",   # typing-only-first-party-import
    "TC002",   # typing-only-third-party-import
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = true

[tool.pytest.ini_options]
testpaths = ["tests"]
